+++ src/RcppExports.cpp
@@ -216,6 +216,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// cpp_triangle
+bool cpp_triangle(arma::mat& D);
+RcppExport SEXP _maotai_cpp_triangle(SEXP DSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat& >::type D(DSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_triangle(D));
+    return rcpp_result_gen;
+END_RCPP
+}
 // eval_gaussian
 double eval_gaussian(arma::vec x, arma::vec mu, arma::mat cov);
 RcppExport SEXP _maotai_eval_gaussian(SEXP xSEXP, SEXP muSEXP, SEXP covSEXP) {
@@ -288,6 +299,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_maotai_integrate_1d", (DL_FUNC) &_maotai_integrate_1d, 2},
     {"_maotai_cpp_pdist", (DL_FUNC) &_maotai_cpp_pdist, 1},
     {"_maotai_cpp_geigen", (DL_FUNC) &_maotai_cpp_geigen, 2},
+    {"_maotai_cpp_triangle", (DL_FUNC) &_maotai_cpp_triangle, 1},
     {"_maotai_eval_gaussian", (DL_FUNC) &_maotai_eval_gaussian, 3},
     {"_maotai_eval_gaussian_data", (DL_FUNC) &_maotai_eval_gaussian_data, 3},
     {"_maotai_eval_gmm_data", (DL_FUNC) &_maotai_eval_gmm_data, 4},